+++ CHANGELOG.md
@@ -10,6 +10,7 @@
  - Add Day to unit conversion
  - Add ability to disable any Input/Output/Function/Widget option (user can view but not change)
  - Add ability to return Input/Output/Function object with module_function()
+ - Add min_value and max_value parameters to query_flux()
 
 
 ## 8.15.8 (2023-04-06)